Update WorldQuestsList.lua
Each world quest is displaying twice globally on the World Quest List display.  They share the same QuestID's but one works and the other doesn't go to the POI.  Have to do a check to see if the QuestID is already displayed and ignore duplicates.

These changes are just a brief example of a temp fix but it isn't perfect.